money_money/requirements_dev.txt
attrs==23.1.0
autobahn
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
logzero==1.7.0
pycparser==2.21
pycrypto==2.6.1
pyotp==2.8.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.31.0
simplejson==3.19.1
six==1.16.0
smartapi-python==1.3.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.3
websocket-client==1.6.0
zope.interface==6.0